+++ game/src/main/kotlin/world/gregs/voidps/world/map/lumbridge/Hans.kts
@@ -1,4 +1,4 @@
-package world.gregs.voidps.world.interact.entity.npc
+package world.gregs.voidps.world.map.lumbridge
 
 import world.gregs.voidps.engine.client.ui.dialogue.Expression
 import world.gregs.voidps.engine.client.ui.dialogue.dialogue
